store/tikv_driver: move gcworker into tikv_driver (#22698)
