+++ store/tikv/interface.go
@@ -19,7 +19,6 @@ import (
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/store/tikv/oracle"
 	"github.com/pingcap/tidb/store/tikv/tikvrpc"
-	"github.com/tikv/pd/client"
 )
 
 // Storage represent the kv.Storage runs on TiKV.
@@ -41,9 +40,6 @@ type Storage interface {
 	// UpdateSPCache updates the cache of safe point.
 	UpdateSPCache(cachedSP uint64, cachedTime time.Time)
 
-	// GetGCHandler gets the GCHandler.
-	GetGCHandler() GCHandler
-
 	// SetOracle sets the Oracle.
 	SetOracle(oracle oracle.Oracle)
 
@@ -56,16 +52,3 @@ type Storage interface {
 	// Closed returns the closed channel.
 	Closed() <-chan struct{}
 }
-
-// GCHandler runs garbage collection job.
-type GCHandler interface {
-	// Start starts the GCHandler.
-	Start()
-
-	// Close closes the GCHandler.
-	Close()
-}
-
-// NewGCHandlerFunc creates a new GCHandler.
-// To enable real GC, we should assign the function to `gcworker.NewGCWorker`.
-var NewGCHandlerFunc func(storage Storage, pdClient pd.Client) (GCHandler, error)
