+++ store/tikv/kv.go
@@ -67,10 +67,8 @@ type KVStore struct {
 	coprCache    *coprCache
 	lockResolver *LockResolver
 	txnLatches   *latch.LatchesScheduler
-	gcWorker     GCHandler
 
-	mock     bool
-	enableGC bool
+	mock bool
 
 	kv        SafePointKV
 	safePoint uint64
@@ -112,7 +110,7 @@ func (s *KVStore) CheckVisibility(startTime uint64) error {
 }
 
 // NewKVStore creates a new TiKV store instance.
-func NewKVStore(uuid string, pdClient pd.Client, spkv SafePointKV, client Client, enableGC bool, coprCacheConfig *config.CoprocessorCache) (*KVStore, error) {
+func NewKVStore(uuid string, pdClient pd.Client, spkv SafePointKV, client Client, coprCacheConfig *config.CoprocessorCache) (*KVStore, error) {
 	o, err := oracles.NewPdOracle(pdClient, time.Duration(oracleUpdateInterval)*time.Millisecond)
 	if err != nil {
 		return nil, errors.Trace(err)
@@ -133,7 +131,6 @@ func NewKVStore(uuid string, pdClient pd.Client, spkv SafePointKV, client Client
 		memCache:        kv.NewCacheDB(),
 	}
 	store.lockResolver = newLockResolver(store)
-	store.enableGC = enableGC
 
 	coprCache, err := newCoprCache(coprCacheConfig)
 	if err != nil {
@@ -155,21 +152,6 @@ func (s *KVStore) EnableTxnLocalLatches(size uint) {
 // IsLatchEnabled is used by mockstore.TestConfig.
 func (s *KVStore) IsLatchEnabled() bool {
 	return s.txnLatches != nil
-}
-
-// StartGCWorker starts GC worker, it's called in BootstrapSession, don't call this function more than once.
-func (s *KVStore) StartGCWorker() error {
-	if !s.enableGC || NewGCHandlerFunc == nil {
-		return nil
-	}
-
-	gcWorker, err := NewGCHandlerFunc(s, s.pdClient)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	gcWorker.Start()
-	s.gcWorker = gcWorker
-	return nil
 }
 
 func (s *KVStore) runSafePointChecker() {
@@ -237,9 +219,6 @@ func (s *KVStore) GetSnapshot(ver kv.Version) kv.Snapshot {
 func (s *KVStore) Close() error {
 	s.oracle.Close()
 	s.pdClient.Close()
-	if s.gcWorker != nil {
-		s.gcWorker.Close()
-	}
 
 	close(s.closed)
 	if err := s.client.Close(); err != nil {
@@ -380,11 +359,6 @@ func (s *KVStore) GetRegionCache() *RegionCache {
 // GetLockResolver returns the lock resolver instance.
 func (s *KVStore) GetLockResolver() *LockResolver {
 	return s.lockResolver
-}
-
-// GetGCHandler returns the GC worker instance.
-func (s *KVStore) GetGCHandler() GCHandler {
-	return s.gcWorker
 }
 
 // Closed returns a channel that indicates if the store is closed.
